+++ src/main/java/com/statoverflow/status/domain/quest/controller/QuestController.java
@@ -4,14 +4,19 @@ import java.util.List;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.statoverflow.status.domain.quest.dto.MainQuestResponseDto;
+import com.statoverflow.status.domain.quest.dto.RerollSubQuestRequestDto;
 import com.statoverflow.status.domain.quest.dto.SubQuestResponseDto;
 import com.statoverflow.status.domain.quest.dto.ThemeResponseDto;
-import com.statoverflow.status.domain.quest.service.QuestService;
+import com.statoverflow.status.domain.quest.service.MainQuestService;
+import com.statoverflow.status.domain.quest.service.SubQuestService;
+import com.statoverflow.status.domain.quest.service.ThemeService;
 import com.statoverflow.status.domain.users.dto.BasicUsersDto;
 import com.statoverflow.status.global.annotation.CurrentUser;
 import com.statoverflow.status.global.response.ApiResponse;
@@ -25,13 +30,15 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class QuestController {
 
-	private final QuestService questService;
+	private final ThemeService themeService;
+	private final MainQuestService mainQuestService;
+	private final SubQuestService subQuestService;
 
 	@GetMapping("/get-themes")
 	public ResponseEntity<ApiResponse<List<ThemeResponseDto>>> getThemes(@CurrentUser BasicUsersDto user,
 		@RequestParam List<Integer> attributes) {
 
-		return ApiResponse.ok(questService.getThemes(attributes));
+		return ApiResponse.ok(themeService.getThemes(attributes));
 
 	}
 
@@ -40,7 +47,7 @@ public class QuestController {
 		@RequestParam List<Integer> attributes,
 		@RequestParam List<Integer> themes) {
 
-		return ApiResponse.ok(questService.rerollThemes(attributes, themes));
+		return ApiResponse.ok(themeService.rerollThemes(attributes, themes));
 
 	}
 
@@ -49,7 +56,7 @@ public class QuestController {
 		@RequestParam List<Integer> attributes,
 		@RequestParam Long theme) {
 
-		return ApiResponse.ok(questService.getMainQuests(attributes, user.id(), theme));
+		return ApiResponse.ok(mainQuestService.getMainQuests(attributes, user.id(), theme));
 
 	}
 
@@ -59,7 +66,7 @@ public class QuestController {
 		@RequestParam Long theme,
 		@RequestParam List<Long> mainQuests) {
 
-		return ApiResponse.ok(questService.rerollMainQuests(attributes, mainQuests, user.id(), theme));
+		return ApiResponse.ok(mainQuestService.rerollMainQuests(attributes, mainQuests, user.id(), theme));
 
 	}
 
@@ -68,7 +75,15 @@ public class QuestController {
 		@RequestParam List<Integer> attributes,
 		@RequestParam Long mainQuest) {
 
-		return ApiResponse.ok(questService.getSubQuests(attributes, mainQuest, user.id()));
+		return ApiResponse.ok(subQuestService.getSubQuests(attributes, mainQuest, user.id()));
+
+	}
+
+	@PostMapping("/reroll-subquests")
+	public ResponseEntity<ApiResponse<List<MainQuestResponseDto>>> rerollSubQuests(@CurrentUser BasicUsersDto user,
+		@RequestBody RerollSubQuestRequestDto dto) {
+		// mainQuestService.subQuestService(dto, user.id());
+		return ApiResponse.ok(null);
 
 	}
