+++ src/main/java/com/statoverflow/status/domain/quest/dto/RerollSubQuestRequestDto.java
@@ -1,0 +1,11 @@
+package com.statoverflow.status.domain.quest.dto;
+
+import java.util.List;
+
+public record RerollSubQuestRequestDto(
+	Long mainQuest,
+	List<Long> attributes,
+	List<Long> selectedSubQuests,
+	List<Long> gottenSubQuests
+) {
+}
