+++ src/main/java/com/statoverflow/status/domain/quest/service/ThemeService.java
@@ -1,0 +1,14 @@
+package com.statoverflow.status.domain.quest.service;
+
+import java.util.List;
+
+import com.statoverflow.status.domain.quest.dto.ThemeResponseDto;
+
+public interface ThemeService {
+
+	List<ThemeResponseDto> getThemes(List<Integer> attributes);
+
+	List<ThemeResponseDto> rerollThemes(List<Integer> attributes, List<Integer> themes);
+
+
+}
